# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This API Handle MELI - Sprint 1 - Grupo 2",
    "title": "MELI Bootcamp API - Sprint 1 - Grupo 2",
    "termsOfService": "https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones",
    "contact": {
      "name": "API Support",
      "url": "https://developers.mercadolibre.com.ar/support"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "paths": {
    "/api/v1/buyers": {
      "get": {
        "description": "Get the details of all buyers on the database.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Buyers"
        ],
        "summary": "List Buyers",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/web_utils.response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/entities.Buyer"
                      }
                    }
                  }
                }
              ]
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          }
        }
      },
      "post": {
        "description": "Save a buyer on the database.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Buyers"
        ],
        "summary": "Create Buyer",
        "parameters": [
          {
            "description": "Buyer to Create",
            "name": "SellerHandler",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.CreateBuyerRequestDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/entities.Buyer"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          }
        }
      }
    },
    "/api/v1/buyers/{id}": {
      "get": {
        "description": "Get the details of a Buyer",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Buyers"
        ],
        "summary": "Get Buyer",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Buyer to be searched",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/entities.Buyer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          }
        }
      },
      "delete": {
        "description": "Delete Buyers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Buyers"
        ],
        "summary": "Delete Buyer",
        "parameters": [
          {
            "type": "string",
            "description": "ID of a Buyer to be excluded",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          }
        }
      },
      "patch": {
        "description": "Update the details of a Buyer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Buyers"
        ],
        "summary": "Update Buyer",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Buyer to be updated",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated Buyer details",
            "name": "Buyer",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.UpdateBuyerRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/entities.Buyer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          }
        }
      }
    },
    "/api/v1/employees": {
      "get": {
        "description": "getAll employees",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Employees"
        ],
        "summary": "List Employees",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "post": {
        "description": "Create employees",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Employees"
        ],
        "summary": "Create Employees",
        "parameters": [
          {
            "description": "EmployeeHandler to Create",
            "name": "Employees",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/entities.RequestCreateEmployee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      }
    },
    "/api/v1/employees/{id}": {
      "get": {
        "description": "Get the details of a Employees",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Employees"
        ],
        "summary": "Get Employees",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Employees to be searched",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "delete": {
        "description": "Delete Employees",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Employees"
        ],
        "summary": "Delete Employees",
        "parameters": [
          {
            "type": "string",
            "description": "ID of a Employees to be excluded",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      },
      "patch": {
        "description": "Update the details of a Employees",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Employees"
        ],
        "summary": "Update Employees",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Employees to be updated",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated Employeesers details",
            "name": "Employees",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/entities.RequestUpdateEmployee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/web_utils.response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "dtos.CreateBuyerRequestDTO": {
      "type": "object",
      "required": [
        "card_number_id",
        "first_name",
        "last_name"
      ],
      "properties": {
        "card_number_id": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        }
      }
    },
...
    "dtos.UpdateBuyerRequestDTO": {
      "type": "object",
      "properties": {
        "card_number_id": {
          "type": "string",
          "x-order": "0"
        },
        "first_name": {
          "type": "string",
          "x-order": "1"
        },
        "last_name": {
          "type": "string",
          "x-order": "2"
        }
      }
    },
...
    "entities.Buyer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "x-order": "0"
        },
        "card_number_id": {
          "type": "string",
          "x-order": "1"
        },
        "first_name": {
          "type": "string",
          "x-order": "2"
        },
        "last_name": {
          "type": "string",
          "x-order": "3"
        }
      }
    },
...
    "entities.RequestCreateEmployee": {
      "type": "object",
      "properties": {
        "card_number_id": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "warehouse_id": {
          "type": "integer"
        }
      }
    },
    "entities.RequestUpdateEmployee": {
      "type": "object",
      "properties": {
        "card_number_id": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "warehouse_id": {
          "type": "integer"
        }
      }
    },
...
    "web_utils.errorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "web_utils.response": {
      "type": "object",
      "properties": {
        "data": {}
      }
    }
  }
}